# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EduSync Bridge API",
        "description": "API de integración bidireccional entre Control Escolar (Tecnológico de Monterrey) y CYPHER Learning NEO LMS v3. Gestiona sincronización de alumnos, docentes, inscripciones, calificaciones y analítica académica.",
        "contact": {
            "email": "admin@edusync.edu"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local"
        }
    ],
    "paths": {
...
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Iniciar sesión",
                "description": "Autentica al usuario. Si el rol requiere 2FA y no está configurado retorna requires_two_factor_setup=true. Si ya está configurado retorna requires_two_factor=true.",
                "operationId": "295c3a0de26c4d16716df38e534dc356",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@edusync.edu"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "Admin@2024!",
                                        "minLength": 8
                                    },
                                    "totp_code": {
                                        "type": "string",
                                        "example": "123456",
                                        "nullable": true,
                                        "maxLength": 6,
                                        "minLength": 6
                                    },
                                    "client_type": {
                                        "description": "WEB=refresh 7 días, MOBILE=30 días",
                                        "type": "string",
                                        "nullable": true,
                                        "default": "WEB",
                                        "enum": [
                                            "WEB",
                                            "MOBILE"
                                        ]
                                    },
                                    "selected_role_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "access_token": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "refresh_token": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "string",
                                                            "format": "uuid"
                                                        },
                                                        "email": {
                                                            "type": "string"
                                                        },
                                                        "first_name": {
                                                            "type": "string"
                                                        },
                                                        "last_name": {
                                                            "type": "string"
                                                        },
                                                        "status": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "active_role": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "string",
                                                            "format": "uuid"
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "super-admin"
                                                        },
                                                        "display_name": {
                                                            "type": "string"
                                                        },
                                                        "hierarchy_level": {
                                                            "type": "integer"
                                                        }
                                                    },
                                                    "type": "object",
                                                    "nullable": true
                                                },
                                                "permissions": {
                                                    "description": "Permisos del rol activo",
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "reports.sync.view"
                                                    }
                                                },
                                                "requires_role_selection": {
                                                    "type": "boolean"
                                                },
                                                "available_roles": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "requires_two_factor": {
                                                    "type": "boolean"
                                                },
                                                "requires_two_factor_setup": {
                                                    "type": "boolean"
                                                },
                                                "must_change_password": {
                                                    "type": "boolean"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "timestamp": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales inválidas"
                    },
                    "423": {
                        "description": "Cuenta bloqueada"
                    },
                    "429": {
                        "description": "Demasiados intentos"
                    }
                }
            }
        },
        "/api/v1/auth/2fa/verify": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Verificar código TOTP (2FA)",
                "description": "Alias de login con totp_code incluido. Completa la autenticación de dos factores.",
                "operationId": "f4a6092628987d207ac973a991198a3e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "totp_code"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "minLength": 8
                                    },
                                    "totp_code": {
                                        "type": "string",
                                        "example": "123456",
                                        "maxLength": 6,
                                        "minLength": 6
                                    },
                                    "client_type": {
                                        "type": "string",
                                        "nullable": true,
                                        "default": "WEB",
                                        "enum": [
                                            "WEB",
                                            "MOBILE"
                                        ]
                                    },
                                    "selected_role_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login exitoso"
                    },
                    "401": {
                        "description": "Credenciales o código TOTP inválidos"
                    },
                    "423": {
                        "description": "Cuenta bloqueada"
                    },
                    "429": {
                        "description": "Demasiados intentos"
                    }
                }
            }
        },
...
        "/api/v1/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Cerrar sesión",
                "description": "Revoca la sesión activa y agrega el access token a la blacklist de Redis.",
                "operationId": "c7f70801e17ec94b9398529425532a95",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Logged out successfully."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}